+++ Changelog.md
@@ -21,6 +21,7 @@ Bug fixes
 
 * Fix operator matchers (`should` syntax) when `method` is redefined on target.
   (Brandon Turner)
+* Fix diffing of hashes with object based keys. (Jon Rowe)
 
 ### 2.14.2 / 2013-08-14
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.14.1...v2.14.2)